feat: Shared `EagerExpr` impl
